moved EntityMapperService to module/folder "model"
